requirements.txt
streamlit==1.32.0
pandas
python-dotenv
playwright==1.46.0
beautifulsoup4==4.12.3
lxml==5.3.0
openai==2.15.0